+++ src/QueryDescription/index.js
@@ -1,9 +1,8 @@
 // @flow
 
-import { propEq, pipe, prop, uniq, map } from 'rambdax'
+import { pipe, prop, uniq, map } from 'rambdax'
 
 // don't import whole `utils` to keep worker size small
-import partition from '../utils/fp/partition'
 import invariant from '../utils/common/invariant'
 import checkName from '../utils/fp/checkName'
 import deepFreeze from '../utils/common/deepFreeze'
@@ -53,8 +52,32 @@ export type On = $RE<{
   left: ColumnName,
   comparison: Comparison,
 }>
-export type Condition = Where | On
-export type QueryDescription = $RE<{ where: Where[], join: On[] }>
+export type SortOrder =
+  | 'asc'
+  | 'desc'
+export const asc: SortOrder = 'asc'
+export const desc: SortOrder = 'desc'
+export type SortBy = $RE<{
+  type: 'sortBy',
+  sortColumn: ColumnName,
+  sortOrder: SortOrder,
+}>
+export type Take = $RE<{
+  type: 'take',
+  count: number,
+}>
+export type Skip = $RE<{
+  type: 'skip',
+  count: number,
+}>
+export type Clause = Where | On | SortBy | Take | Skip
+export type QueryDescription = $RE<{
+  where: Where[],
+  join: On[],
+  sortBy: SortBy[],
+  take: ?Take,
+  skip: ?Skip,
+}>
 
 const columnSymbol = Symbol('Q.column')
 const comparisonSymbol = Symbol('QueryComparison')
@@ -222,6 +245,20 @@ export function or(...conditions: Where[]): Or {
   return { type: 'or', conditions }
 }
 
+function sortBy(sortColumn: ColumnName, sortOrder: SortOrder = asc): SortBy {
+  return { type: 'sortBy', sortColumn, sortOrder }
+}
+
+function take(count: number): Take {
+  return { type: 'take', count }
+}
+
+function skip(count: number): Skip {
+  return { type: 'skip', count }
+}
+
+export { sortBy as experimentalSortBy, take as experimentalTake, skip as experimentalSkip }
+
 // Note: we have to write out three separate meanings of OnFunction because of a Babel bug
 // (it will remove the parentheses, changing the meaning of the flow type)
 type _OnFunctionColumnValue = (TableName<any>, ColumnName, Value) => On
@@ -255,7 +292,31 @@ export const on: OnFunction = (table, leftOrWhereDescription, valueOrComparison)
 }
 
 const syncStatusColumn = columnName('_status')
-const getJoins: (Condition[]) => [On[], Where[]] = (partition(propEq('type', 'on')): any)
+const extractClauses: (Clause[]) => QueryDescription = clauses => {
+  const clauseMap = { join: [], sortBy: [], where: [], take: null, skip: null }
+  clauses.forEach(cond => {
+    let { type } = cond
+    switch (type) {
+      case 'take':
+      case 'skip':
+        // $FlowFixMe: Flow is too dumb to realize that it is valid
+        clauseMap[type] = cond
+        break
+      default:
+      case 'where':
+        type = 'where'
+        // fallthrough
+      case 'on':
+        if (type === 'on') type = 'join'
+        // fallthrough
+      case 'sortBy':
+        clauseMap[type].push(cond)
+        break
+    }
+  })
+  // $FlowFixMe: Flow is too dumb to realize that it is valid
+  return clauseMap
+}
 const whereNotDeleted = where(syncStatusColumn, notEq('deleted'))
 const joinsWithoutDeleted = pipe(
   map(prop('table')),
@@ -263,10 +324,12 @@ const joinsWithoutDeleted = pipe(
   map(table => on(table, syncStatusColumn, notEq('deleted'))),
 )
 
-export function buildQueryDescription(conditions: Condition[]): QueryDescription {
-  const [join, whereConditions] = getJoins(conditions)
+export function buildQueryDescription(clauses: Clause[]): QueryDescription {
+  const clauseMap = extractClauses(clauses)
 
-  const query = { join, where: whereConditions }
+  invariant(!(clauseMap.skip && !clauseMap.take), 'cannot skip without take')
+
+  const query = clauseMap
   if (process.env.NODE_ENV !== 'production') {
     deepFreeze(query)
   }
@@ -277,6 +340,7 @@ export function queryWithoutDeleted(query: QueryDescription): QueryDescription {
   const { join, where: whereConditions } = query
 
   const newQuery = {
+    ...query,
     join: [...join, ...joinsWithoutDeleted(join)],
     where: [...whereConditions, whereNotDeleted],
   }